Refactor WIP - Gestures inside Zowi
