+++ lib/zowi.js
@@ -14,7 +14,24 @@ module.exports = () => {
         };
     };
 
+    const gestures = {
+        'ZowiHappy': 'H 1',
+        'ZowiSuperHappy': 'H 2',
+        'ZowiSad': 'H 3',
+        'ZowiSleeping': 'H 4',
+        'ZowiFart': 'H 5',
+        'ZowiConfused': 'H 6',
+        'ZowiLove': 'H 7',
+        'ZowiAngry': 'H 8',
+        'ZowiFretful': 'H 9',
+        'ZowiMagic': 'H 10',
+        'ZowiWave': 'H 11',
+        'ZowiVictory': 'H 12',
+        'ZowiFail': 'H 13'
+    };
+
     return {
         start,
+        gestures,
     };
 };
